Show previously hidden navigation when screen > 850px"
When the screen starts off lower than 850px, or when the user hides the navigation, it (i.e. the  #main-menu element) is hidden by jQuery. When widening the screen to above 850px, the navigation, if hidden, remained hidden. This required changes in both CSS and JS to solve.

I added some additonal rules to the .hidden class as applied to the navigation (i.e. #main-menu.hidden), when the screen exceeds 850px.

This was not originally being applied, however, because jQuery was
leaving behind an inline style on the #main-menu element. This could
have been solved by using the !important override in CSS, but since it is
considered harmful -- and it is my own jQuery call that caused the issue
in the first place -- I decided to solve it in JS.

After performing the jQuery animate() on #main-menu, I implemented a
'complete' function that toggles the .hidden class and removes the
inline style.

I also had to change the initial hiding of #main-menu in JS to toggle
the class instead of using the hide() method.